notes-backend/requirements.txt
fastapi>=0.70.0
uvicorn>=0.15.0
motor>=2.5.1
pydantic[email]>=1.8.2
pydantic-settings>=2.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.5
python-dotenv>=0.19.2
email-validator>=1.1.3 